Revert "feat: add swap details screen"
This reverts commit 667c318d21eec3c9cbb0681ed1b342bef7225d69.
