+++ src/swap/SwapTransactionDetails.tsx
@@ -12,15 +12,12 @@ import TokenDisplay, { formatValueToDisplay } from 'src/components/TokenDisplay'
 import Touchable from 'src/components/Touchable'
 import InfoIcon from 'src/icons/InfoIcon'
 import { getLocalCurrencySymbol, usdToLocalCurrencyRateSelector } from 'src/localCurrency/selectors'
-import { navigate } from 'src/navigator/NavigationService'
-import { Screens } from 'src/navigator/Screens'
 import { useSelector } from 'src/redux/hooks'
 import colors from 'src/styles/colors'
 import { typeScale } from 'src/styles/fonts'
 import { Spacing } from 'src/styles/styles'
 import ExchangeRateIcon from 'src/swap/icons/ExchangeRateIcon'
 import FeesIcon from 'src/swap/icons/FeesIcon'
-import { QuoteResult } from 'src/swap/types'
 import { useTokenInfo } from 'src/tokens/hooks'
 import { TokenBalance } from 'src/tokens/slice'
 import { getTokenId } from 'src/tokens/utils'
@@ -42,7 +39,6 @@ interface Props {
     percentage: BigNumber
   }
   estimatedCrossChainFee?: BigNumber
-  quote: QuoteResult | null
 }
 
 interface PropsOld extends Props {
@@ -401,7 +397,6 @@ export function SwapTransactionDetails({
   feeTokenId: networkFeeTokenId,
   estimatedCrossChainFee,
   fetchingSwapQuote: loading,
-  quote,
 }: Props) {
   const { t } = useTranslation()
   const localCurrencySymbol = useSelector(getLocalCurrencySymbol)
@@ -413,13 +408,7 @@ export function SwapTransactionDetails({
   })
 
   const handleShowMoreDetails = () => {
-    if (quote && toToken && fromToken) {
-      navigate(Screens.SwapDetailsScreen, {
-        toTokenId: toToken?.tokenId,
-        fromTokenId: fromToken?.tokenId,
-        quote,
-      })
-    }
+    // TODO: show more details
   }
 
   if (!fromToken || !toToken || !exchangeRatePrice) {
